requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.3
djangorestframework==3.16.0
Faker==37.4.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
fonttools==4.58.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
MouseInfo==0.1.3
mysqlclient==2.2.7
numpy==2.3.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
psycopg2==2.9.10
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyparsing==3.2.3
pyperclip==1.9.0
pypng==0.20220715.0
PyQRCode==1.2.1
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
requests==2.32.4
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wikipedia==1.4.0
